Remove the mockbuild target
The target has always been defined in rpkg and here, originally with
identical code but recently a change in rpkg had not been made here as
well, so it was lost to the user.

Also, the target was only registered here, not in rpkg, but this recently
changed.

All in all, the target should be registered and implemented in rpkg, not
here.